apps/backend/requirements.txt
# AI Agent Team Backend - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Storage
supabase==2.0.3
postgrest==0.13.1

# Redis/Queue
redis==5.0.1

# HTTP Client
httpx==0.25.2

# LLM Integration
openai==1.3.7  # For OpenRouter API (compatible interface)

# Security
pyjwt==2.8.0
cryptography==41.0.7

# Email (Gmail API)
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Web Scraping & Search
duckduckgo-search==4.1.1  # FREE web search, no API key needed
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Retry Logic
tenacity==8.2.3

# Data Processing
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# YAML (for TOON converter)
pyyaml==6.0.1

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
